+++ test/test_helper.exs
@@ -1,4 +1,5 @@
 PhoenixFlags.TestRepo.start_link()
+PhoenixFlags.TestEndpoint.start_link()
 
 Ecto.Adapters.SQL.Sandbox.mode(PhoenixFlags.TestRepo, :manual)